Fix security level of account selecting in database (uint8 instead of uint32)

+++ src/server/game/Server/WorldSocket.cpp
@@ -1106,7 +1106,7 @@ int WorldSocket::HandleAuthSession(WorldPacket& recvPacket)
     else
     {
         fields = result->Fetch();
-        security = fields[1].GetInt32();
+        security = fields[1].GetUInt8();
     }
 
     // Re-check account ban (same check as in realmd)
